Update README.md
This is just a a brief run down of what I'm trying to do especially offline at localhost because we always know as much as we take it for granted that this is all online it's always better to do thing offline then go live for a specific task, such as updating rather than just installing by default users should be able to pick out what specific packages they want in their distro or operating system. So who cares I don't because it always works for me, at least I make it work but this can be irritating when it involves countless man hours to get something to work in the first place because the internet doesn't work as consistently as required by apps rather users, a kind of worm picked up from running smartphones and tablets requiring gigabytes of constant traffic. This is not what each device is there for.